+++ config/manager/manager.yaml
@@ -90,9 +90,9 @@ spec:
             - name: RELATED_IMAGE_internal_rest_backup_server
               value: quay.io/eclipse/che-backup-server-rest:eeacd92
             - name: RELATED_IMAGE_gateway_authentication_sidecar
-              value: quay.io/openshift/origin-oauth-proxy:4.7
+              value: quay.io/oauth2-proxy/oauth2-proxy:v7.2.0
             - name: RELATED_IMAGE_gateway_authorization_sidecar
-              value: quay.io/openshift/origin-kube-rbac-proxy:4.7
+              value: quay.io/brancz/kube-rbac-proxy:v0.11.0
             - name: RELATED_IMAGE_gateway_header_sidecar
               value: quay.io/che-incubator/header-rewrite-proxy:latest
             - name: CHE_FLAVOR
