+++ pkg/util/util.go
@@ -480,8 +480,7 @@ func NewBoolPointer(value bool) *bool {
 }
 
 func IsNativeUserModeEnabled(c *orgv1.CheCluster) bool {
-	// Native user mode is now available only on openshift
-	return IsOpenShift && c.Spec.Auth.NativeUserMode != nil && *c.Spec.Auth.NativeUserMode
+	return c.Spec.Auth.NativeUserMode != nil && *c.Spec.Auth.NativeUserMode
 }
 
 // IsOAuthEnabled returns true when oAuth is enable for CheCluster resource, otherwise false.
